refactor: split network tests into seperate files
